+++ pkg/core/actions/actions_test.go
@@ -128,48 +128,3 @@ var testSha = m.DockerImageSha("sha1")
 var testImage = m.Image{Name: "image1", Sha: testSha}
 var testCont = m.Container{Image: testImage}
 var testPod = m.Pod{Namespace: "abc", Name: "def", UID: "fff", Containers: []m.Container{testCont}}
-
-func TestAddPodAction(t *testing.T) {
-	// actual
-	actual := m.NewModel(&m.Config{}, "test version")
-	(&AddPod{testPod}).Apply(actual)
-	// expected (a bit hacky to get the times set up):
-	//  - pod gets added to .Pods
-	//  - all images within pod get added to .Images
-	//  - all new images get added to hub check queue
-	expected := *m.NewModel(&m.Config{}, "test version")
-	expected.Pods[testPod.QualifiedName()] = testPod
-	imageInfo := m.NewImageInfo(testSha, "image1")
-	imageInfo.ScanStatus = m.ScanStatusInHubCheckQueue
-	imageInfo.TimeOfLastStatusChange = actual.Images[testSha].TimeOfLastStatusChange
-	expected.Images[testSha] = imageInfo
-	expected.ImageHubCheckQueue = append(expected.ImageHubCheckQueue, imageInfo.ImageSha)
-	//
-	assertEqual(t, actual, expected)
-}
-
-func TestAddImageAction(t *testing.T) {
-	// actual
-	actual := m.NewModel(&m.Config{ConcurrentScanLimit: 3}, "test version")
-	(&AddImage{testImage}).Apply(actual)
-	// expected (a bit hacky to get the times set up):
-	//  - image gets added to .Images
-	//  - image gets added to hub check queue
-	expected := *m.NewModel(&m.Config{ConcurrentScanLimit: 3}, "test version")
-	imageInfo := m.NewImageInfo(testSha, "image1")
-	imageInfo.ScanStatus = m.ScanStatusInHubCheckQueue
-	imageInfo.TimeOfLastStatusChange = actual.Images[testSha].TimeOfLastStatusChange
-	expected.Images[testSha] = imageInfo
-	expected.ImageHubCheckQueue = append(expected.ImageHubCheckQueue, imageInfo.ImageSha)
-	//
-	assertEqual(t, actual, expected)
-}
-
-// AllPods does remove pre-existing pods
-func TestAllPods(t *testing.T) {
-	actual := createNewModel1()
-	(&AllPods{}).Apply(actual)
-	if len(actual.Pods) != 0 {
-		t.Errorf("expected 0 pods, found %d", len(actual.Pods))
-	}
-}
